Switch-style recoding of values with string pattern matching

Switch-style recoding of values with string pattern matching

grep_case(x, ..., preserve = FALSE, default = NA)

Arguments

x
A vector

...
<dynamic-dots> A sequence of two-sided formulas. Elements of x that match the left hand side (LHS) will be replaced with the value in the right hand side (RHS).

The LHS must evaluate to a character string.

The RHS must be of length one.

NULL inputs are ignored.

preserve
If TRUE, unmatched elements of x will be returned unmodified. (The elements may have their type coerced to be compatible with replacement values.) If FALSE, unmatched elements of x will be replaced with default. Defaults to FALSE.

default
If preserve is FALSE, a value to replace unmatched elements of x. Defaults to NA.

Value

A vector of the same length as x.

See also

Examples

words <- c("caterpillar", "dogwood", "catastrophe", "dogma")

grep_case(
  words,
  "cat" ~ "feline",
  "dog" ~ "canine"
)

#> [1] "feline" "canine" "feline" "canine"

caps_words <- c("caterpillar", "dogwood", "Catastrophe", "DOGMA")

grep_case(
  caps_words,
  "cat" ~ "feline",
  "dog" ~ "canine",
  ignore.case = TRUE
)

#> [1] "feline" "canine" "feline" "canine"